review user

+++ OnlineShop/Controllers/ReviewsController.cs
@@ -1,4 +1,5 @@
-﻿using OnlineShop.Models;
+﻿using Microsoft.AspNet.Identity;
+using OnlineShop.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -22,6 +23,7 @@ namespace OnlineShop.Controllers
         {
             try
             {
+                rev.UserId = User.Identity.GetUserId();
                 db.Reviews.Add(rev);
                 db.SaveChanges();
                 return Redirect("/Products/Show/" + rev.ProductId);
@@ -37,7 +39,16 @@ namespace OnlineShop.Controllers
         public ActionResult Edit(int id)
         {
             Review rev = db.Reviews.Find(id);
-            return View(rev);
+
+            if (rev.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                return View(rev);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                return RedirectToAction("Index", "Products");
+            }
         }
 
         [HttpPut]
@@ -47,17 +58,19 @@ namespace OnlineShop.Controllers
             try
             {
                 Review rev = db.Reviews.Find(id);
-                if (TryUpdateModel(rev))
-                {
-                    rev.ReviewComment = requestRev.ReviewComment;
-                    rev.ReviewRating = requestRev.ReviewRating;
-                    db.SaveChanges();
-                }
+                if(rev.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+                    if (TryUpdateModel(rev))
+                    {
+                        rev.ReviewComment = requestRev.ReviewComment;
+                        rev.ReviewRating = requestRev.ReviewRating;
+                        db.SaveChanges();
+                    }
                 return Redirect("/Products/Show/" + rev.ProductId);
             }
             catch (Exception e)
             {
-                return View();
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                return RedirectToAction("Index", "Products");
             }
 
         }
@@ -66,9 +79,17 @@ namespace OnlineShop.Controllers
         public ActionResult Delete(int id)
         {
             Review rev = db.Reviews.Find(id);
-            db.Reviews.Remove(rev);
-            db.SaveChanges();
-            return Redirect("/Products/Show/" + rev.ProductId);
+            if (rev.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                db.Reviews.Remove(rev);
+                db.SaveChanges();
+                return Redirect("/Products/Show/" + rev.ProductId);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                return RedirectToAction("Index", "Products");
+            }
         }
 
     }
